added content in pgae.jsx
