+++ Next-Assignment/src/app/assignment-2/page.jsx
@@ -1,12 +1,11 @@
-'use-client'
-import React from 'react'
+import React from 'react';
 
 const Page = () => {
   return (
-    <div>
-    
+    <div style={{ textAlign: 'center' }}>
+      <h1>Assignment-2</h1>
     </div>
-  )
-}
+  );
+};
 
-export default Page
+export default Page;
