--- Next-Assignment/src/app/assignment-3/layout.jsx
@@ -1,18 +1,0 @@
-'use client'
-import SidebarDrawer from "@/component/SideBar";
-// import Sidebar from "@/component/SideBar";
-import { AuthProvider } from "@/context/AuthProvider";
-
-export default function RootLayout({ children }) {
-  const questionArray = [1, 2, 3, 4, 5, 6];
-
-  return (
-    <AuthProvider>
-      <div className="assignment-layout">
-        {/* <Sidebar countArray={questionArray} assignmentNo="3" /> */}
-        <SidebarDrawer countArray={questionArray} assignmentNo="3" />
-        <main className="main-content">{children}</main>
-      </div>
-    </AuthProvider>
-  );
-}
