+++ Next-Assignment/src/app/assignment-3/page.jsx
@@ -1,9 +1,11 @@
-"use client";
-
-import React from "react";
+import React from 'react';
 
 const Page = () => {
-  return <div></div>;
+  return (
+    <div style={{ textAlign: 'center' }}>
+      <h1>Assignment-3</h1>
+    </div>
+  );
 };
 
 export default Page;
