+++ Next-Assignment/src/app/assignment-4/page.jsx
@@ -1,11 +1,11 @@
-import React from 'react'
-
+import React from 'react';
 
 const Page = () => {
   return (
-    <div>
+    <div style={{ textAlign: 'center' }}>
+      <h1>Assignment-4</h1>
     </div>
-  )
-}
+  );
+};
 
-export default Page
+export default Page;
